requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
GitPython>=3.1.40
requests>=2.31.0
